Remove `GameTickManager` and related clock handling, enhance `Metronome` with tick catchup limits and disposal logic, improve `RealmWorld` and `RealmZone` lifecycle and lazy loading mechanics, add `ActorRegistrar` for actor management, and expand `WorldComposer` tile management logic.

+++ FluffyByte.Realm.Game/Entities/Actors/Components/StatBase.cs
@@ -10,7 +10,7 @@ namespace FluffyByte.Realm.Game.Entities.Actors.Components;
 
 public class StatBase
 {
-    public byte Base { get; set; } = 10;
+    public int Base { get; set; } = 10;
     public int Equipment { get; set; } = 0;
     public int Temporary { get; set; } = 0;
